extract location methods into a shared concern
