+++ app/models/spot.js
@@ -1,47 +1,21 @@
 import DS from 'ember-data';
+import Locatable from './concerns/locatable';
 
-export default DS.Model.extend({
+var Spot = DS.Model.extend(Locatable, {
   game: DS.belongsTo('game', { async: true }),
   user: DS.belongsTo('user', { async: true }),
   guesses: DS.hasMany('guess', { async: true }),
-
   imageUrl: DS.attr(),
-  location: DS.attr(),
   createdAt: DS.attr('date'),
   userName: DS.attr(),
   order: function() {
     return parseInt(this.get('id'), 10);
   }.property('id'),
-
   formattedCreatedAt: function() {
     return moment(this.get('createdAt')).fromNow();
   }.property('createdAt'),
 
-  longitude: function() {
-    if( this.get('location') === undefined || this.get('location') === null) {
-      return null;
-    } else {
-      return this.get('location')['coordinates'][0];
-    }
-  }.property('location'),
-
-  latitude: function() {
-    if( this.get('location') === undefined || this.get('location') === null) {
-      return null;
-    } else {
-      return this.get('location')['coordinates'][1];
-    }
-  }.property('location'),
-
-  formattedLocation: function() {
-    var latitude = this.get('latitude');
-    var longitude = this.get('longitude');
-    if( latitude === undefined || longitude === undefined) {
-      return "missing";
-    } else {
-      var latLabel = (latitude > 0) ? "N" : "S";
-      var lonLabel = (longitude > 0) ? "E" : "W";
-      return Math.abs(latitude) + "°" + latLabel + " " + Math.abs(longitude) + "°" + lonLabel;
-    }
-  }.property('latitude', 'longitude')
 });
+
+
+export default Spot;
